fix page size to include footers in bottom margin
The margins for the NSM guidelines *include* the footer where the page
numbers are. By default, LaTeX does not measure this part when setting
up margins.

The margins are set to be a little over the 1 inch minimum so that there
is no doubt that it fits the requirements.

This removes the settings from `uhthesis3.sty`.